Fixed type bug

+++ server/heartbeat.go
@@ -221,7 +221,7 @@ func processNewMembershipList(newMembership *MembershipList) {
 			sort.Strings(Membership.List)
 
 		// If the new Membership has a more recent time, update it
-		} else if pingTime < newPingTime && math.Abs(pingTime) != newPingTime {
+		} else if pingTime < newPingTime && int64(math.Abs(float64(pingTime))) != newPingTime {
 			Membership.Data[nextHostname] = newPingTime
 
 			// If the hostname is not in the list but it's in the data map,
